remove unsed files from ion-sfu
